+++ src/v0/form/js/main.js
@@ -917,7 +917,7 @@ $('.cr-military-status').slideUp();
   /**
    * modal functions
    */
-  $('#privacy').on('click', function(){
+  /*$('#privacy').on('click', function(){
     $('<div class="cr-modal"><div class="cr-modal-bg">&nbsp;</div><div class="cr-modal-offset"><div class="cr-modal-content"><div class="cr-modal-inner"><h2>Privacy act statement</h2><p>The information you provide will permit the Consumer Financial Protection Bureau to respond to your complaint or inquiry about companies and services we supervise. Information about your complaint or inquiry (including your personally identifiable information) may be shared:</p><p>• with the entity that is the subject of your complaint;</p><p>• with third parties as necessary to get information relevant to resolving a complaint;</p><p>• with a court, a party in litigation, a magistrate, an adjudicative body or administrative tribunal in the course of a proceeding, or the Department of Justice;</p><p>• with other federal or state agencies or regulatory authorities for enforcement and statutory purposes; and</p><p>• with contractors, agents, and others authorized by the CFPB to receive this information.</p><p>We may also share your complaint or inquiry (but not your personally identifiable information) with the public through a public complaint database.</p><p>This collection of information is authorized by 12 U.S.C. § 5493.</p><p>You are not required to file a complaint or share any identifying information, including your Social Security number, and you may withdraw your complaint at any time. However, if you do not include the requested information, the CFPB may not be able to act on your complaint.</p><a href="#" class="cr-modal-close">Close</a></div></div></div></div>').appendTo('body');
     $(".cr-modal").animate({
       opacity: 1
@@ -925,28 +925,30 @@ $('.cr-military-status').slideUp();
   });
 
   $('#omb').on('click', function(){
-/*     e.preventDefault(); */
+    // e.preventDefault();
     $('<div class="cr-modal"><div class="cr-modal-bg">&nbsp;</div><div class="cr-modal-offset"><div class="cr-modal-content"><div class="cr-modal-inner"><h2>OMB #3170-0011</h2><p>An agency may not conduct or sponsor, and a person is not required to respond to, a collection of information unless the collection of information displays a valid control number assigned by the Office of Management and Budget (OMB). The OMB control number for this collection is 3170-0011, expires 11/30/2014.</p><a href="#" class="cr-modal-close">Close</a></div></div></div></div>').appendTo('body');
     $(".cr-modal").animate({
       opacity: 1
     }, 600 );
   });
+*/
 
-
+/*
   $('#narrative-consent-modal').on('click', function(){
     $('<div class="cr-modal"> <div class="cr-modal-bg">&nbsp;</div> <div class="cr-modal-offset"> <div class="cr-modal-content"> <div class="cr-modal-inner"> <h2>Include your description of what happened on consumerfinance.gov</h2> <p>When you submit a complaint through the CFPB, you can choose to publish your description of what happened in the Consumer Complaint Database on our website, consumerfinance.gov. This will help researchers, journalists, and other consumers understand the financial marketplace though reading your first-hand account.</p> <p>You should avoid including personal information, such as names, addresses, and account numbers in the content of your description, but we will also conduct a series of automated and manual reviews to remove remaining personal information before publishing. The following example highlights information that would be removed:</p> <p id="scrubbed-narrative">We took out a private education loan for our daughter <span class="redacted-text">XXXXXX</span> while she was attending the <span class="redacted-text">XXXXXX</span> University in <span class="redacted-text">XXXXXX</span>, MA. This loan went into default after she transferred to a community college in <span class="redacted-text">XXXXXX</span>, MA. Our lender received notice that she was no longer a full-time student in <span class="redacted-text">XX</span>/<span class="redacted-text">XX</span>/2011. Then our younger son <span class="redacted-text">XXXXXXX</span> became very sick and we could no longer make the monthly payments on our daughter’s loan. We tried to set up a new payment plan to get us back on track, but the servicer won’t negotiate. We\'ve been paying what we can but we still haven\'t reduced any of the principal on the loan. </p>  <p>We monitor this process to ensure that all personal information is removed, but a small risk remains that something will be missed. We won’t publish your description until we have finished the process of removing personal information, so your description won’t appear in the database right away.</p> <p>Publishing your description on consumerfinance.gov is completely optional. If you authorize publication but change your mind, you can withdraw your authorization at any time by calling the CFPB help line at (855) 411-2372. Your description will then be removed the next time the database is updated.</p><p>Whether you choose to provide or withdraw your authorization will not affect how your complaint is handled.</p> <a href="#" class="cr-modal-close">Close</a></div></div></div></div>').appendTo('body');
     $(".cr-modal").animate({
       opacity: 1
     }, 600 );
   });
+*/
 
-
+/*
   $('#save-and-continue-later-link').on('click', function(){
     $('<div class="cr-modal"> <div class="cr-modal-bg">&nbsp;</div> <div class="cr-modal-offset"> <div class="cr-modal-content"> <div class="cr-modal-inner"> <h2>Create an account so you can continue later</h2> <p>Log in or create an account and we will send you a link and a complaint ID number that you can use to finish submitting this complaint.</p> <div id="save-options" class="select-product-or-issue select-save-method">                          <label class="radio block save-options-label">                 <input name="radio_" type="radio" class="radio_input product save-options-radio save-with-email selected" value="save-with-email" />                  Create an account                 <br /><small></small>             </label>                                                                            <label class="radio block save-options-label">                 <input name="radio_" type="radio" class="radio_input save-options-radio product save-with-login" value="save-with-login" />                  Log in                 <br /><small></small>             </label> </div> <hr class="save-method-hr" /> <fieldset id="save-with-email" class="save-method">     <div class="span3 cr-label cr-save-option email">       <label for="con1-email">      Email address or Phone number       <div class="is-required">Answer Required</div>      </label>      <input type="email" name="cr-email2" id="cr-email2_2356852120719850" placeholder="" class="input-large left">       <p>We will send a link to this email address to continue this complaint</p>     </div>    <div class="span3 cr-label cr-save-option">       <label for="con1-email">      Create a password       <div class="is-required">Answer Required</div>      </label>      <input type="email" name="cr-password" id="cr-email2_2356852120719850" placeholder="" class="input-large left">       <p>Include letters and numbers, at least 8 characters</p>     </div>     </fieldset> <fieldset id="save-with-phone" class="save-method">        <div class="span3 cr-label cr-save-option">       <label for="con1-email">      Phone number        <div class="is-required">Answer Required</div>      </label>      <input type="email" name="cr-email2" id="cr-email2_2356852120719850" placeholder="" class="input-large left">     <p>We will send a text message to this phone number. Standard messaging rates may apply.</p>    </div>    <div class="span3 cr-label cr-save-option">       <label for="con1-email">      Create a password       <div class="is-required">Answer Required</div>      </label>      <input type="email" name="cr-password" id="cr-email2_2356852120719850" placeholder="" class="input-large left">       <p>Include letters and numbers, at least 8 characters</p>     </div>     </fieldset> <fieldset id="save-with-login" class="save-method">    <p>If you have submitted a complaint to the CFPB in the past, log in to save this complaint to your account. </p>         <div class="span3 cr-label cr-save-option email">       <label for="con1-email">      Email address or phone number       <div class="is-required">Answer Required</div>      </label>      <input type="email" name="cr-email2" id="cr-email2_2356852120719850" placeholder="" class="input-large left">     </div>    <div class="span3 cr-label cr-save-option">       <label for="con1-email">      Password      <div class="is-required">Answer Required</div>      </label>      <input type="email" name="cr-password" id="cr-email2_2356852120719850" placeholder="" class="input-large left">           </div>             </fieldset> <a id="save-and-continue-with-email" class="save-with-email save-type" href="#"><button class="btn" type="button">Create an account and save progress</button></a> <a id="save-and-continue-with-phone" class="save-with-phone save-type" href="#"><button class="btn" type="button">Save progress</button></a> <a id="save-and-continue-with-login" class="save-with-login save-type" href="#"><button class="btn" type="button">Log in and save progress</button></a><p>&nbsp;</p> <a href="#" class="cr-modal-close">Close</a></div></div></div></div>').appendTo('body');
     $(".cr-modal").animate({
       opacity: 1
     }, 600 );
-/*     $('input.save-with-email').value(':checked'); */
+//     $('input.save-with-email').value(':checked');
     $('input.save-with-email').parent().addClass('active');
     $('input.save-with-email').prop("checked", true);
   $('fieldset#save-with-phone').hide();
@@ -983,7 +985,7 @@ $('.cr-military-status').slideUp();
   });
 
   });
-
+*/
   $('#save-and-continue-review-page').on('click', function(){
     $('<div class="cr-modal"><div class="cr-modal-bg">&nbsp;</div><div class="cr-modal-offset"><div class="cr-modal-content"><div class="cr-modal-inner"><h2>Complaint saved!</h2> <p class="your-id-number">Your complaint draft ID is <strong>141030-43002</strong></p> <p>We have sent an email to <strong>p.robinson@gmail.com</strong> that includes this complaint draft ID and a link to log in to your account and finish submitting your complaint.</p> <p>You can also complete this complaint over the phone by calling us at (855) 411-2372 and providing this draft ID number along with your email address and phone number.</p> <p>Make sure to come back soon! Your link will expire on November 10th, 2014. After that date, your information will be securely removed from our servers.</p> <a href="#" class="cr-modal-close">Close</a></div></div></div></div>').appendTo('body');
     $(".cr-modal").animate({
@@ -991,6 +993,7 @@ $('.cr-military-status').slideUp();
     }, 600 );
   });
 
+/*
   $('body').on('click', '.cr-modal-close', function(){
     $(".cr-modal").animate({
       opacity: 0
@@ -1005,5 +1008,5 @@ $('.cr-military-status').slideUp();
       $('.cr-modal').remove();
     });
   });
-
+*/
   // show/hide save options